actualizacion de persona

+++ src/main/java/com/example/backendportafolio/Dto/DtoPerson.java
@@ -18,7 +18,7 @@ public class DtoPerson {
 
     private String email;
 
-    private Integer phone;
+    private Long phone;
 
     private String nationality;
 
@@ -34,7 +34,7 @@ public class DtoPerson {
     public DtoPerson() {
     }
 
-    public DtoPerson(String name, String lastName, String birth, String title, String email, Integer phone, String nationality, String aboutMe, String bannerUrl, String profileUrl) {
+    public DtoPerson(String name, String lastName, String birth, String title, String email, Long phone, String nationality, String aboutMe, String bannerUrl, String profileUrl) {
         this.name = name;
         this.lastName = lastName;
         this.birth = birth;
